# Refactor tiling code generation