--- a/Deeploy/CommonExtensions/CodeTransformationPasses/IntrospectiveCodeTransformation.py
+++ b/Deeploy/CommonExtensions/CodeTransformationPasses/IntrospectiveCodeTransformation.py
@@ -23,16 +23,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import copy
 import types
 from typing import Dict, List
 
 import mako.codegen as codegen
 from mako.lexer import Lexer
-from mako.parsetree import Expression, TemplateNode
+from mako.parsetree import Expression, TemplateNode, Text
+from mako.template import Template
 
 from Deeploy.AbstractDataTypes import Pointer, Struct
-from Deeploy.DeeployTypes import ExecutionBlock, NetworkContext, NodeTemplate, OperatorRepresentation, VariableBuffer
+from Deeploy.DeeployTypes import ExecutionBlock, NetworkContext, OperatorRepresentation, VariableBuffer
 
 _NULL: str = "NULL"
 
@@ -42,65 +42,76 @@ class IntrospectiveCodeTransformationMixIn():
     parseTreeDict: Dict[int, TemplateNode] = {}
 
     @staticmethod
-    def _generateParseTree(template: NodeTemplate) -> TemplateNode:
-        return Lexer(template.template._source).parse()
+    def _generateParseTree(template: Template) -> TemplateNode:
+        return Lexer(template._source).parse()
 
     @staticmethod
-    def _reconstructCode(template: NodeTemplate, node: TemplateNode):
-
-        def fixupParseTree(parseTree: TemplateNode) -> TemplateNode:
-            nodes = []
-            prevLine = 0
-            prevPos = 0
-            for node in parseTree.nodes:
-
-                newNode = copy.copy(node)
-                offset = len(node.source)
-
-                # Expression contain the actual expression + the symbols "${}", i.e. 3 offset symbols
-                if isinstance(newNode, Expression):
-                    offset += 3
-
-                prevPos = prevPos + offset
-
-                if prevLine != node.lineno:
-                    prevPos = node.pos
-
-                newNode.pos = prevPos
-                prevLine = node.lineno
-
-                nodes.append(newNode)
-
-            parseTree.nodes = nodes
-
-            return parseTree
-
-        node = fixupParseTree(node)
-
-        temp = template.template
-        lexer = Lexer(temp._source)
+    def _reconstructCode(template: Template, node: TemplateNode) -> Template:
+        lexer = Lexer(template._source)
         source = codegen.compile(
             node,
-            temp.uri,
+            template.uri,
             None,
-            default_filters = temp.default_filters,
-            buffer_filters = temp.buffer_filters,
-            imports = temp.imports,
-            future_imports = temp.future_imports,
+            default_filters = template.default_filters,
+            buffer_filters = template.buffer_filters,
+            imports = template.imports,
+            future_imports = template.future_imports,
             source_encoding = lexer.encoding,
             generate_magic_comment = True,
-            strict_undefined = temp.strict_undefined,
-            enable_loop = temp.enable_loop,
-            reserved_names = temp.reserved_names,
+            strict_undefined = template.strict_undefined,
+            enable_loop = template.enable_loop,
+            reserved_names = template.reserved_names,
         )
-        module = types.ModuleType(temp.module_id)
-        code = compile(source, temp.module_id, "exec")
+        module = types.ModuleType(template.module_id)
+        code = compile(source, template.module_id, "exec")
         exec(code, module.__dict__, module.__dict__)
 
-        temp._code = code
-        temp.module = module
-        temp.callable_ = temp.module.render_body
-        template.template = temp
+        template._code = code
+        template.module = module
+        template.callable_ = template.module.render_body
+        return template
+
+    @staticmethod
+    def _indexPointer(parseTree: TemplateNode, ptrName: str, index: str) -> TemplateNode:
+        indexes = [i for i, node in enumerate(parseTree.nodes) if isinstance(node, Expression) and node.text == ptrName]
+
+        for offset, idx in enumerate(indexes):
+            bracketOpen = Text("[", source = "[", lineno = 0, pos = 0, filename = None)
+            indexExpr = Expression(index, '', source = index, lineno = 0, pos = 0, filename = None)
+            bracketClose = Text("]", source = "]", lineno = 0, pos = 0, filename = None)
+            parseTree.nodes.insert(idx + 3 * offset + 1, bracketOpen)
+            parseTree.nodes.insert(idx + 3 * offset + 2, indexExpr)
+            parseTree.nodes.insert(idx + 3 * offset + 3, bracketClose)
+
+        return parseTree
+
+    @staticmethod
+    def indexVars(template: Template, varNames: List[str], index: str) -> None:
+        if len(varNames) == 0:
+            return
+        parseTree = IntrospectiveCodeTransformationMixIn._generateParseTree(template)
+        for name in varNames:
+            parseTree = IntrospectiveCodeTransformationMixIn._indexPointer(parseTree, name, index)
+        IntrospectiveCodeTransformationMixIn._reconstructCode(template, parseTree)
+
+    @staticmethod
+    def _dereferencePointer(parseTree: TemplateNode, ptrName: str) -> TemplateNode:
+        indexes = [i for i, node in enumerate(parseTree.nodes) if isinstance(node, Expression) and node.text == ptrName]
+
+        for offset, idx in enumerate(indexes):
+            text = Text("*", source = "*", lineno = 0, pos = 0, filename = None)
+            parseTree.nodes.insert(idx + offset, text)
+
+        return parseTree
+
+    @staticmethod
+    def dereferenceVars(template: Template, varNames: List[str]) -> None:
+        if len(varNames) == 0:
+            return
+        parseTree = IntrospectiveCodeTransformationMixIn._generateParseTree(template)
+        for name in varNames:
+            parseTree = IntrospectiveCodeTransformationMixIn._dereferencePointer(parseTree, name)
+        IntrospectiveCodeTransformationMixIn._reconstructCode(template, parseTree)
 
     def extractDynamicReferences(self,
                                  ctxt: NetworkContext,
@@ -112,7 +123,7 @@ class IntrospectiveCodeTransformationMixIn():
         for codeSnippet in executionBlock.codeSnippets:
             template, operatorRepresentation = codeSnippet.template, codeSnippet.operatorRepresentation
 
-            newRefs = self._extractDynamicExpressions(ctxt, operatorRepresentation, template, unrollStructs,
+            newRefs = self._extractDynamicExpressions(ctxt, operatorRepresentation, template.template, unrollStructs,
                                                       includeGobalReferences)
 
             makoDynamicReferences += newRefs
@@ -132,11 +143,10 @@ class IntrospectiveCodeTransformationMixIn():
     def _extractDynamicExpressions(self,
                                    ctxt: NetworkContext,
                                    operatorRepresentation: OperatorRepresentation,
-                                   template: NodeTemplate,
+                                   template: Template,
                                    unrollStructs = False,
                                    includeGobalReferences = False):
-
-        codeHash = hash(template.template._source)
+        codeHash = hash(template._source)
 
         if codeHash in self.parseTreeDict.keys():
             makoParseTree = self.parseTreeDict[codeHash]
@@ -146,60 +156,43 @@ class IntrospectiveCodeTransformationMixIn():
             self.parseTreeDict[codeHash] = makoParseTree
 
         # Filter parsing tree for expressions
-        makoExpressions = [node.text for node in makoParseTree.nodes if type(node) == Expression]
+        makoExpressions = [node.text for node in makoParseTree.nodes if isinstance(node, Expression)]
+
+        # Filter represented expressions
+        representedExpressions = [
+            operatorRepresentation[expr] for expr in makoExpressions if expr in operatorRepresentation
+        ]
+
+        # Filter buffers from expressions
+        references = [expr for expr in representedExpressions if ctxt.is_buffer(expr)]
+
+        if unrollStructs:
+
+            def _unrollStructReferences(val: Struct) -> List[str]:
+                assert isinstance(val, Struct)
+                # Recursively unroll struct references
+                structReferences = []
+                for field in val.value.values():
+                    if isinstance(field, Struct):
+                        structReferences += _unrollStructReferences(field)
+                    elif isinstance(field, Pointer) and field.referenceName != _NULL:
+                        structReferences.append(field.referenceName)
+                return structReferences
+
+            # Unroll local struct references
+            for ref in references:
+                if hasattr(ctxt.lookup(ref), "structDict"):
+                    references += _unrollStructReferences(ctxt.lookup(ref).structDict)
 
         # Filter expressions for local variables contained in operatorRepresentation
-        makoLocalReferences = [
-            node for node in makoExpressions
-            if ((node in operatorRepresentation) and type(operatorRepresentation[node]) == str and (
-                operatorRepresentation[node] in ctxt.localObjects.keys()))
-        ]
+        localReferences = [ref for ref in references if ctxt.is_local(ref)]
 
         # Filter expressions for global variables contained in operatorRepresentation
-        makoGlobalReferences = [
-            node for node in makoExpressions
-            if ((node in operatorRepresentation) and type(operatorRepresentation[node]) == str and (
-                operatorRepresentation[node] in ctxt.globalObjects.keys()))
-        ]
-
-        def _unrollStructReferences(val) -> List[str]:
-            # Unroll struct references
-            structReferences = []
-            if isinstance(val, Struct):
-                for key, _type in val.value.items():
-                    if isinstance(_type, Struct):
-                        structReferences += _unrollStructReferences(val.value[key])
-                    elif isinstance(_type, Pointer) and val.value[key].referenceName != _NULL:
-                        structReferences.append(val.value[key].referenceName)
-            return structReferences
-
-        # Unroll local struct references
-        localReferences = []
-        localStructReferences = []
-        for ref in makoLocalReferences:
-            localReferences.append(operatorRepresentation[ref])
-            if unrollStructs:
-                if ctxt.is_local(operatorRepresentation[ref]) and hasattr(ctxt.lookup(operatorRepresentation[ref]),
-                                                                          "structDict"):
-                    localStructReferences += _unrollStructReferences(
-                        ctxt.lookup(operatorRepresentation[ref]).structDict)
-
-        # Unroll global struct references
-        globalReferences = []
-        globalStructReferences = []
-        for ref in makoGlobalReferences:
-            globalReferences.append(operatorRepresentation[ref])
-            if unrollStructs:
-                if ctxt.is_global(operatorRepresentation[ref]) and hasattr(ctxt.lookup(operatorRepresentation[ref]),
-                                                                           "structDict"):
-                    globalStructReferences += _unrollStructReferences(
-                        ctxt.lookup(operatorRepresentation[ref]).structDict)
+        globalReferences = [ref for ref in references if ctxt.is_global(ref)]
 
         # Filter for dynamically allocated tensors
-        dynamicLocalReferences = [ref for ref in localReferences + localStructReferences if ctxt.lookup(ref)._deploy]
-        dynamicGlobalReferences = [
-            ref for ref in globalReferences + globalStructReferences if isinstance(ctxt.lookup(ref), VariableBuffer)
-        ]
+        dynamicLocalReferences = [ref for ref in localReferences if ctxt.lookup(ref)._deploy]
+        dynamicGlobalReferences = [ref for ref in globalReferences if isinstance(ctxt.lookup(ref), VariableBuffer)]
 
         if includeGobalReferences:
             return dynamicLocalReferences + dynamicGlobalReferences