Main contribution of this PR is abstracting the DMA logic into the `AsyncDma` class and showing it works with the currently supported DMAs like the PULP's mchan and L3 dma, and Snitch's cluster dma.
The main goal was enabling any-dimensional transfers which got incarnated as the `AnydimAsyncDmaTransferAdapter`.

## Performance comparison

<details>

<summary>Comparison of network execution on Siracusa emulated in GVSoC</summary>

| Network                | Conf  | L1 size | Cycles before | Cycles after | Diff     | Perc. diff |
| ---------------------- | ----- | ------- | ------------- | ------------ | -------- | ---------- |
| Simple regression      | SB L2 | 64k     | 636738        | 637758       | 1020     | 0.2%       |
| Simple regression      | SB L3 | 64k     | 1007791       | 989804       | \-17987  | \-1.8%     |
| Simple regression      | DB L2 | 64k     | 648051        | 643213       | \-4838   | \-0.7%     |
| Simple regression      | DB L3 | 64k     | 954903        | 988656       | 33753    | 3.5%       |
| MobileNetV2            | SB L3 | 64k     | Not runnable  | 198539808    | #VALUE!  | #VALUE!    |
| MobileNetV2            | DB L3 | 64k     | Not runnable  | 194710802    | #VALUE!  | #VALUE!    |
| miniMobileNetV2        | SB L2 | 16k     | 120580        | 110003       | \-10577  | \-8.8%     |
| miniMobileNetV2        | SB L3 | 16k     | 392763        | 361312       | \-31451  | \-8.0%     |
| miniMobileNetV2        | DB L2 | 16k     | 122368        | 110997       | \-11371  | \-9.3%     |
| miniMobileNetV2        | DB L3 | 16k     | 362411        | 361283       | \-1128   | \-0.3%     |
| microLlama/microLlama1 | SB L2 | 10k     | 657015        | 519142       | \-137873 | \-21.0%    |
| microLlama/microLlama1 | SB L3 | 10k     | 4461938       | 3977350      | \-484588 | \-10.9%    |
| microLlama/microLlama1 | DB L2 | 10k     | 707464        | 542643       | \-164821 | \-23.3%    |
| microLlama/microLlama1 | DB L3 | 10k     | 3913082       | 4047015      | 133933   | 3.4%       |
| CCT/CCT_1_16_16_8      | SB L2 | 64k     | 493469        | 485747       | \-7722   | \-1.6%     |
| CCT/CCT_1_16_16_8      | SB L3 | 64k     | 1255862       | 1219939      | \-35923  | \-2.9%     |
| CCT/CCT_1_16_16_8      | DB L2 | 64k     | 513180        | 504340       | \-8840   | \-1.7%     |
| CCT/CCT_1_16_16_8      | DB L3 | 64k     | 1227566       | 1213998      | \-13568  | \-1.1%     |

</details>

I wanted to test also Snitch but the program only prints 0 cycles. Visually checking the code, there is no fundamental change except for the fact that we now emit less barriers.

## Added
- AsyncDma abstraction of DMA's
- test runner per DMA and a script that tests all the DMA's
- generic Single/DoubleBufferingTilingCodeGeneration classes
- TilingVariableReplacementUpdate class that updates the variable replacement refs
- TilingHoistingMixIn class that encapsulates all the hoisting helper functions of tiling
- sorting of input memory allocations to allow references that live in the same memory level as the memory they are referencing
- a function that tests the tiling solution for correctness which currently only tests buffer allocation for byte alignment
- IntrospectiveCodeTransformation: `_indexPointer()`, `indexVars()`, `dereferenceVars()`. The `*Vars` functions index/dereference a list of variables (useful for tiling)
- NetworkContext: `unravelReference()` that unravels a `_ReferenceBuffer` until the base buffer
- NetworkContext: `is_object()` - helper function that determines whether the string represents a name of a local or global object
- NetworkContext: `is_buffer()` - helper function that determines whether the string represents a name of a buffer
- missing checks for environment variables
- `_permuteHyperRectangle` helper function

## Changed
- mchan HAL is now reduced to bare-bones
- refactor of the IntrospectiveCodeTransformation to work on the Mako template and made imo clearer
- refactor of memory allocation code transformation passes
- _ReferenceBuffer accepts an optional `offset` argument to offset the reference
- NetworkContext: `hoistReference` - accepts the actual buffer as reference instead of name, accepts shape, offset, and override_type arguments, and returns the actual buffer, not its name
- `_mangleNodeRep` -> `_mangleOpRepr` - the canonical way we use is `OperatorRepresentation`. `NodeRep` and `ParseDict` are old iterations of that.
- rename of permutation functions to follow this convention: `permute` is an action that permutes something, `permutation` is a function that generates a permutation
- `_permuteList` to just `_permute`
- removed manual buffer name mangling since we do it in the ExecutionBlock generate() function, simplifies templates
- we now check that buffer shapes/hyperrectangles/tiling ranks match which required changing a few `serializeTilingSolution` functions to preserve the same shape rank
- big refactor of the code generation part of the TilingExtension
- port of PULPOpen tiling code generation
- port of Snitch tiling code generation
- PULPClusterTilingSB and PULPClusterTilingDB now allow for transfers of any rank (dimensionality)
- PULP's final output diff is now calculated as absolute error, instead of just subtraction
- common code generation code between testMVP/generateNetwork/... was extracted into a single `generateTestNetwork` function
- in some functions, instead of passing the name of a buffer, the actual buffer is just passed
- tile function allows overriding the optimizer with external tilingSolution and memoryMap
- refactor of the permutation functions for clarity


## PR Merge Checklist

1. [x] The PR is rebased on the latest `devel` commit and pointing to `devel`.
2. [x] Your PR reviewed and approved.
3. [x] All checks are passing.
4. [x] The `CHANGELOG.md` file has been updated.
5. [x] If the docker was modified, change back its link after review.